+++ websiteapp/__init__.py
@@ -9,5 +9,7 @@
 # @license   https://www.opensource.org/licenses/mit-license.php MIT
 # @link      https://github.com/MarcinOrlowski/website-as-app
 #
+# @file      websiteapp/const.py
+#
 #####
 """